pyproject.toml
[project]
name = "opengenes-mcp"
version = "0.1.0"
description = "MCP for OpenGenes"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "eliot>=1.17.5",
    "pycomfort>=0.0.18",
    "fastmcp>=2.6.1",
    "fastapi>=0.115.12",
    "aiosqlite>=0.20.0",
    "sqlite-utils>=3.37"
]

[project.scripts]
opengenes-mcp = "opengenes_mcp.server:cli_app_stdio"
server = "opengenes_mcp.server:cli_app"
stdio = "opengenes_mcp.server:cli_app_stdio"
sse = "opengenes_mcp.server:cli_app_sse"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opengenes_mcp"]

[tool.hatch.build]
include = [
    "src/",
    "data/"
]
exclude = [
    "*.log",
    "logs/",
    "images/",
    ".venv/",
    ".pytest_cache/",
    "test/",
    # Exclude development/temporary databases but keep the main data
    "*.db",
    "temp*.sqlite*",
    "test*.sqlite*",
    "debug*.sqlite*"
]

[dependency-groups]
dev = [
    "just-agents-core>=0.7.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
